requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
tzdata==2024.2
Django
djangorestframework
django-cors-headers
Pillow
rest_framework_simplejwt
decouple

# Twilio package
twilio  # Specify your preferred version range